# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amo Market Portal API",
        "description": "This is the Portal API documentation for the Amo Market project.",
        "contact": {
            "email": "amomarket.info@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/portal/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all products based on current location",
                "description": "Returns a list of products grouped by category, available within a 2 km radius from the user's current location.",
                "operationId": "getAllProduct",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Current latitude of the user",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Current longitude of the user",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products Retrieved Successfully."
                                        },
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "sub_category_id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Milk"
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "example": "https://example.com/image.png"
                                                            },
                                                            "weight": {
                                                                "type": "string",
                                                                "example": "1L"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 50
                                                            },
                                                            "original_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 60
                                                            },
                                                            "discount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 10
                                                            },
                                                            "delivery_time": {
                                                                "type": "string",
                                                                "example": "20 mins"
                                                            },
                                                            "status": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "property": "product"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Current location not provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Current location not provided."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Products List Not Found in the radius",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Products List Not Found in the radius."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}